+++ cmd/frontend/graphqlbackend/codemod_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/search/query"
+	"github.com/sourcegraph/sourcegraph/internal/search/query"
 )
 
 func TestCodemod_validateArgsNoRegex(t *testing.T) {
