+++ cmd/frontend/graphqlbackend/search_didyoumeanquoted_test.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/search/query"
+	"github.com/sourcegraph/sourcegraph/internal/search/query"
 
 	"github.com/davecgh/go-spew/spew"
 )
