+++ internal/search/query/pattern_type.go
