+++ internal/search/query/pattern_type_test.go
