+++ internal/search/query/syntax/NOTICE
