+++ internal/search/query/syntax/parser_test.go
