+++ internal/search/query/syntax/scanner.go
