+++ internal/search/query/syntax/scanner_test.go
