+++ internal/search/query/syntax/tokentype_string.go
