+++ internal/search/search/search.go
@@ -4,9 +4,9 @@ package search
 import (
 	"regexp/syntax"
 
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/pkg/search/query"
 	"github.com/sourcegraph/sourcegraph/internal/endpoint"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
+	"github.com/sourcegraph/sourcegraph/internal/search/query"
 )
 
 // PatternInfo is the struct used by vscode pass on search queries. Keep it in
